requirements.txt
transformers==4.57.5
Flask==3.1.2
flask-cors==6.0.2
gunicorn==23.0.0
openai==2.15.0
python-dotenv==1.2.1
requests==2.32.4
nltk==3.9.2
PyPDF2==3.0.1
fonttools==4.59.0